ib_simulator/requirements.txt
# Interactive Brokers API Simulator Requirements

# Database
duckdb>=0.9.2

# Configuration
pyyaml>=6.0

# Server framework
asyncio  # Built-in, but listing for clarity

# Data handling
numpy>=1.21.0
pandas>=1.5.0
scipy>=1.9.0

# Date/time handling
python-dateutil>=2.8.2

# Password hashing
bcrypt>=4.0.1

# Logging
colorlog>=6.7.0

# Market data simulation
yfinance>=0.2.12  # For realistic market data patterns

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Type hints
typing-extensions>=4.8.0